+++ include/corner_detection.hpp
@@ -13,7 +13,7 @@ void readSelection(const cv::Mat& img_in, std::vector<cv::Point2f>& corners);
 /**
  * Detect automatically the arena corners
 */
-std::vector<cv::Point2f> autodetect(const cv::Mat& img_in){
+std::vector<cv::Point2f> autodetect(const cv::Mat& img_in) {
     std::vector<cv::Point2f> corners;
 
     // convert to grayscale
@@ -35,7 +35,7 @@ std::vector<cv::Point2f> autodetect(const cv::Mat& img_in){
         cv::Mat drawing = cv::Mat::zeros( mask.size(), CV_8UC3 ); //Todo: REMOVE
     #endif
 
-    for( int i = 0; i< contours.size(); i++ ){
+    for (size_t i = 0; i < contours.size(); i++) {
         float ctArea= cv::contourArea(contours[i]);
         if(ctArea > biggestContourArea)
         {
@@ -90,7 +90,7 @@ std::vector<cv::Point2f> autodetect(const cv::Mat& img_in){
 /**
  * Read the configuration, if not existent ask the user to select points
 */
-std::vector<cv::Point2f> manualSelect(const cv::Mat& img_in, std::string config_folder){
+std::vector<cv::Point2f> manualSelect(const cv::Mat& img_in, std::string config_folder) {
     //
     //  Read configuration file or ask for manual selection
     //
@@ -149,11 +149,11 @@ cv::Point findLineCenter(const cv::Mat& img_in, const std::vector<cv::Point> &ar
     cv::Mat maskRect(size, CV_8U, cv::Scalar(0.0));
 
     std::vector<cv::Point2f> tempcorners;
-    for(int i = 0; i < 4; ++i)
+    for (int i = 0; i < 4; ++i)
         tempcorners.emplace_back(arena[i].x, arena[i].y);
     cv::Scalar white = cv::Scalar(255.0); // white
     cv::Point vertices[4];
-    for(int i = 0; i < 4; ++i){
+    for (int i = 0; i < 4; ++i) {
         vertices[i] = tempcorners[i];
     }
     cv::fillConvexPoly(maskRect, vertices, 4, white);
@@ -170,7 +170,7 @@ cv::Point findLineCenter(const cv::Mat& img_in, const std::vector<cv::Point> &ar
     int biggestContourIdxForLine = -1;
     float biggestContourAreaForLine = 0;
 
-    for( int i = 0; i< contoursForLine.size(); i++ ){
+    for (size_t i = 0; i< contoursForLine.size(); i++) {
         float ctArea= cv::contourArea(contoursForLine[i]);
         if(ctArea > biggestContourAreaForLine)
         {
@@ -183,7 +183,7 @@ cv::Point findLineCenter(const cv::Mat& img_in, const std::vector<cv::Point> &ar
 
     int xsum = 0;
     int ysum = 0;
-    for(int i = 0; i < 4; i++){
+    for (int i = 0; i < 4; i++) {
         xsum += approx_curve_forLine[i].x;
         ysum += approx_curve_forLine[i].y;
     }
@@ -214,7 +214,7 @@ void readSelection(const cv::Mat& img_in, std::vector<cv::Point2f>& corners) {
     while (points.size() < 4) {
         cv::imshow(windowname, img_in);
 
-        for (int i=0; i < points.size(); i++) {
+        for (size_t i=0; i < points.size(); i++) {
             cv::circle(img_in, points[i], 20, cv::Scalar(240,0,0),CV_FILLED);
         }
         cv::waitKey(1);
