+++ include/polygon_utils.hpp
@@ -82,7 +82,7 @@ namespace PUtils {
         bool swapped = true;
         while (swapped) {
             swapped = false;
-            for (int i = 0; i< polygon.size()-1; ++i) {
+            for (size_t i = 0; i< polygon.size()-1; ++i) {
                 if (pointOrderComparison(polygon[i],polygon[i+1],polyCenter)) {
                     //Swap
                     Point temp = polygon[i];
